Final version for Week3 project
This is part of the Coursera Capstone Project, IBM.  
BeautifulSoup is used to pull data of Toronto Postcode. 
Next, saving it in CSV file. 
Then, cleaning and preparing the libraries. 
After that, Foursquare API was used to get JSON and location, i.e. latitude and longitude.
K-means was used for clustering.
Toronto map was shown the markers of cluster labels.